requirements.txt
PyYaml==6.0.1
firebase-admin==6.2.0
pip==23.2.1
pytz==2023.3
requests==2.31.0
ruff==0.0.290
setuptools==68.2.2